pyproject.toml
[tool.poetry]
name = "llmrag"
version = "0.1.0"
description = ""
authors = ["pakkio <claudio.pacchiega@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
pymupdf = "^1.23.28"
sentence-transformers = "^2.2.2"
torch = "^2.1.0"
numpy = "^1.24.0"
chromadb = "^0.4.22"
scikit-learn = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
